Got rid of unit test .db files from the repos; they're created in the temp directory now
Moved node cache and utility code into separate source files

Finished first draft of node code, getting unit tests to pass is next

+++ tests/test_utils.go
@@ -4,6 +4,7 @@ import (
 	"database/sql"
 	"fmt"
 	"os"
+	"runtime/debug"
 	"testing"
 
 	_ "github.com/mattn/go-sqlite3"
@@ -12,10 +13,16 @@ import (
 )
 
 func GetTestDb(name string) *sql.DB {
+	tmp_file, tmp_err := os.CreateTemp(os.TempDir(), name)
+	if tmp_err != nil {
+		panic(fmt.Sprintf("GetTestDb fails CreateTemp: %v", tmp_err))
+	}
+	name = tmp_file.Name()
+
 	os.Remove(name)
 	db, err := sql.Open("sqlite3", name)
 	if err != nil {
-		panic(fmt.Sprintf("GetTestDb fails: %v", err))
+		panic(fmt.Sprintf("GetTestDb fails Open: %v", err))
 	}
 	gonedb.Setup(db)
 	return db
@@ -23,28 +30,38 @@ func GetTestDb(name string) *sql.DB {
 
 func AssertEqual[T comparable](t *testing.T, expected T, got T) {
 	if got != expected {
+		debug.PrintStack()
 		t.Errorf("AssertEqual: expected %v - got %v", expected, got)
-		t.Fail()
+		t.Fatal()
 	}
 }
 
 func AssertTrue(t *testing.T, check bool) {
 	if !check {
+		debug.PrintStack()
 		t.Errorf("AssertTrue fails")
-		t.Fail()
+		t.Fatal()
 	}
 }
 
 func AssertError(t *testing.T, err error) {
 	if err == nil {
+		debug.PrintStack()
 		t.Errorf("AssertError: %v", err)
-		t.Fail()
+		t.Fatal()
 	}
 }
 
 func AssertNoError(t *testing.T, err error) {
 	if err != nil {
+		debug.PrintStack()
 		t.Errorf("AssertNoError: %v", err)
-		t.Fail()
+		t.Fatal()
 	}
+}
+
+func GetStringId(t *testing.T, db *sql.DB, val string) int64 {
+	id, err := gonedb.Strings.GetId(db, val)
+	AssertNoError(t, err)
+	return id
 }
